Split RunTest and BytecodeTest into parts, put in matching packages.
