+++ test/junit/scala/lang/primitives/PredefAutoboxingTest.scala
@@ -1,11 +1,9 @@
-package scala
+package scala.lang.primitives
 
-import org.junit.Test
 import org.junit.Assert._
+import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
-
-import scala.tools.testing.AssertUtil._
 
 @RunWith(classOf[JUnit4])
 class PredefAutoboxingTest {
