+++ test/junit/scala/lang/traits/BytecodeTest.scala
@@ -1,4 +1,4 @@
-package scala.issues
+package scala.lang.traits
 
 import org.junit.Assert._
 import org.junit.Test
@@ -9,7 +9,6 @@ import scala.collection.JavaConverters._
 import scala.tools.asm.Opcodes
 import scala.tools.asm.Opcodes._
 import scala.tools.asm.tree.ClassNode
-import scala.tools.nsc.backend.jvm.AsmUtils
 import scala.tools.partest.ASMConverters._
 import scala.tools.testing.BytecodeTesting
 import scala.tools.testing.BytecodeTesting._
@@ -17,194 +16,6 @@ import scala.tools.testing.BytecodeTesting._
 @RunWith(classOf[JUnit4])
 class BytecodeTest extends BytecodeTesting {
   import compiler._
-
-  @Test
-  def t8731(): Unit = {
-    val code =
-      """class C {
-        |  def f(x: Int) = (x: @annotation.switch) match {
-        |    case 1 => 0
-        |    case 2 => 1
-        |    case 3 => 2
-        |  }
-        |  final val K = 10
-        |  def g(x: Int) = (x: @annotation.switch) match {
-        |    case K => 0
-        |    case 1 => 10
-        |    case 2 => 20
-        |  }
-        |}
-      """.stripMargin
-
-    val c = compileClass(code)
-
-    assertTrue(getInstructions(c, "f").count(_.isInstanceOf[TableSwitch]) == 1)
-    assertTrue(getInstructions(c, "g").count(_.isInstanceOf[LookupSwitch]) == 1)
-  }
-
-  @Test
-  def t8926(): Unit = {
-    import scala.reflect.internal.util.BatchSourceFile
-
-    // this test cannot be implemented using partest because of its mixed-mode compilation strategy:
-    // partest first compiles all files with scalac, then the java files, and then again the scala
-    // using the output classpath. this shadows the bug SI-8926.
-
-    val annotA =
-      """import java.lang.annotation.Retention;
-        |import java.lang.annotation.RetentionPolicy;
-        |@Retention(RetentionPolicy.RUNTIME)
-        |public @interface AnnotA { }
-      """.stripMargin
-    val annotB = "public @interface AnnotB { }"
-
-    val scalaSrc =
-      """@AnnotA class A
-        |@AnnotB class B
-      """.stripMargin
-
-    val run = new global.Run()
-    run.compileSources(List(new BatchSourceFile("AnnotA.java", annotA), new BatchSourceFile("AnnotB.java", annotB), new BatchSourceFile("Test.scala", scalaSrc)))
-    val outDir = global.settings.outputDirs.getSingleOutput.get
-    val outfiles = (for (f <- outDir.iterator if !f.isDirectory) yield (f.name, f.toByteArray)).toList
-
-    def check(classfile: String, annotName: String) = {
-      val f = (outfiles collect { case (`classfile`, bytes) => AsmUtils.readClass(bytes) }).head
-      val descs = f.visibleAnnotations.asScala.map(_.desc).toList
-      assertTrue(descs.toString, descs exists (_ contains annotName))
-    }
-
-    check("A.class", "AnnotA")
-
-    // known issue SI-8928: the visibility of AnnotB should be CLASS, but annotation classes without
-    // a @Retention annotation are currently emitted as RUNTIME.
-    check("B.class", "AnnotB")
-  }
-
-  @Test
-  def t6288bJumpPosition(): Unit = {
-    val code =
-      """object Case3 {                                 // 01
-        | def unapply(z: Any): Option[Int] = Some(-1)   // 02
-        | def main(args: Array[String]) {               // 03
-        |    ("": Any) match {                          // 04
-        |      case x : String =>                       // 05
-        |        println("case 0")                      // 06 println and jump at 6
-        |      case _ =>                                // 07
-        |        println("default")                     // 08 println and jump at 8
-        |    }                                          // 09
-        |    println("done")                            // 10
-        |  }
-        |}
-      """.stripMargin
-    val List(mirror, module) = compileClasses(code)
-
-    val unapplyLineNumbers = getInstructions(module, "unapply").filter(_.isInstanceOf[LineNumber])
-    assert(unapplyLineNumbers == List(LineNumber(2, Label(0))), unapplyLineNumbers)
-
-    val expected = List(
-      LineNumber(4, Label(0)),
-      LineNumber(5, Label(5)),
-      Jump(IFEQ, Label(20)),
-
-      LineNumber(6, Label(11)),
-      Invoke(INVOKEVIRTUAL, "scala/Predef$", "println", "(Ljava/lang/Object;)V", false),
-      Jump(GOTO, Label(33)),
-
-      LineNumber(5, Label(20)),
-      Jump(GOTO, Label(24)),
-
-      LineNumber(8, Label(24)),
-      Invoke(INVOKEVIRTUAL, "scala/Predef$", "println", "(Ljava/lang/Object;)V", false),
-      Jump(GOTO, Label(33)),
-
-      LineNumber(10, Label(33)),
-      Invoke(INVOKEVIRTUAL, "scala/Predef$", "println", "(Ljava/lang/Object;)V", false)
-    )
-
-    val mainIns = getInstructions(module, "main") filter {
-      case _: LineNumber | _: Invoke | _: Jump => true
-      case _ => false
-    }
-    assertSameCode(mainIns, expected)
-  }
-
-  @Test
-  def bytecodeForBranches(): Unit = {
-    val code =
-      """class C {
-        |  def t1(b: Boolean) = if (b) 1 else 2
-        |  def t2(x: Int) = if (x == 393) 1 else 2
-        |  def t3(a: Array[String], b: AnyRef) = a != b && b == a
-        |  def t4(a: AnyRef) = a == null || null != a
-        |  def t5(a: AnyRef) = (a eq null) || (null ne a)
-        |  def t6(a: Int, b: Boolean) = if ((a == 10) && b || a != 1) 1 else 2
-        |  def t7(a: AnyRef, b: AnyRef) = a == b
-        |  def t8(a: AnyRef) = Nil == a || "" != a
-        |}
-      """.stripMargin
-
-    val c = compileClass(code)
-
-    // t1: no unnecessary GOTOs
-    assertSameCode(getMethod(c, "t1"), List(
-      VarOp(ILOAD, 1), Jump(IFEQ, Label(6)),
-      Op(ICONST_1), Jump(GOTO, Label(9)),
-      Label(6), Op(ICONST_2),
-      Label(9), Op(IRETURN)))
-
-    // t2: no unnecessary GOTOs
-    assertSameCode(getMethod(c, "t2"), List(
-      VarOp(ILOAD, 1), IntOp(SIPUSH, 393), Jump(IF_ICMPNE, Label(7)),
-      Op(ICONST_1), Jump(GOTO, Label(10)),
-      Label(7), Op(ICONST_2),
-      Label(10), Op(IRETURN)))
-
-    // t3: Array == is translated to reference equality, AnyRef == to null checks and equals
-    assertSameCode(getMethod(c, "t3"), List(
-      // Array ==
-      VarOp(ALOAD, 1), VarOp(ALOAD, 2), Jump(IF_ACMPEQ, Label(23)),
-      // AnyRef ==
-      VarOp(ALOAD, 2), VarOp(ALOAD, 1), VarOp(ASTORE, 3), Op(DUP), Jump(IFNONNULL, Label(14)),
-      Op(POP), VarOp(ALOAD, 3), Jump(IFNULL, Label(19)), Jump(GOTO, Label(23)),
-      Label(14), VarOp(ALOAD, 3), Invoke(INVOKEVIRTUAL, "java/lang/Object", "equals", "(Ljava/lang/Object;)Z", false), Jump(IFEQ, Label(23)),
-      Label(19), Op(ICONST_1), Jump(GOTO, Label(26)),
-      Label(23), Op(ICONST_0),
-      Label(26), Op(IRETURN)))
-
-    val t4t5 = List(
-      VarOp(ALOAD, 1), Jump(IFNULL, Label(6)),
-      VarOp(ALOAD, 1), Jump(IFNULL, Label(10)),
-      Label(6), Op(ICONST_1), Jump(GOTO, Label(13)),
-      Label(10), Op(ICONST_0),
-      Label(13), Op(IRETURN))
-
-    // t4: one side is known null, so just a null check on the other
-    assertSameCode(getMethod(c, "t4"), t4t5)
-
-    // t5: one side known null, so just a null check on the other
-    assertSameCode(getMethod(c, "t5"), t4t5)
-
-    // t6: no unnecessary GOTOs
-    assertSameCode(getMethod(c, "t6"), List(
-      VarOp(ILOAD, 1), IntOp(BIPUSH, 10), Jump(IF_ICMPNE, Label(7)),
-      VarOp(ILOAD, 2), Jump(IFNE, Label(12)),
-      Label(7), VarOp(ILOAD, 1), Op(ICONST_1), Jump(IF_ICMPEQ, Label(16)),
-      Label(12), Op(ICONST_1), Jump(GOTO, Label(19)),
-      Label(16), Op(ICONST_2),
-      Label(19), Op(IRETURN)))
-
-    // t7: universal equality
-    assertInvoke(getMethod(c, "t7"), "scala/runtime/BoxesRunTime", "equals")
-
-    // t8: no null checks invoking equals on modules and constants
-    assertSameCode(getMethod(c, "t8"), List(
-      Field(GETSTATIC, "scala/collection/immutable/Nil$", "MODULE$", "Lscala/collection/immutable/Nil$;"), VarOp(ALOAD, 1), Invoke(INVOKEVIRTUAL, "java/lang/Object", "equals", "(Ljava/lang/Object;)Z", false), Jump(IFNE, Label(10)),
-      Ldc(LDC, ""), VarOp(ALOAD, 1), Invoke(INVOKEVIRTUAL, "java/lang/Object", "equals", "(Ljava/lang/Object;)Z", false), Jump(IFNE, Label(14)),
-      Label(10), Op(ICONST_1), Jump(GOTO, Label(17)),
-      Label(14), Op(ICONST_0),
-      Label(17), Op(IRETURN)))
-  }
 
   def checkForwarder(classes: Map[String, ClassNode], clsName: Symbol, target: String) = {
     val List(f) = getMethods(classes(clsName.name), "f")
@@ -417,36 +228,37 @@ class BytecodeTest extends BytecodeTesting {
       // pass `null` to super constructor, no box-unbox, no Integer created
       List(ALOAD, ILOAD, PUTFIELD, ALOAD, ACONST_NULL, "<init>", RETURN))
   }
+
 }
 
 object invocationReceiversTestCode {
   // if cloneType is more specific than Object (e.g., String), a bridge method is generated.
   def definitions(cloneType: String) =
-   s"""trait T { override def clone(): $cloneType = "hi" }
-      |trait U extends T
-      |class C1 extends U with Cloneable {
-      |  // The comments below are true when $cloneType is Object.
-      |  // C1 gets a forwarder for clone that invokes T.clone. this is needed because JVM method
-      |  // resolution always prefers class members, so it would resolve to Object.clone, even if
-      |  // C1 is a subtype of the interface T which has an overriding default method for clone.
-      |
-      |  // invokeinterface T.clone
-      |  def f1 = (this: T).clone()
-      |
-      |  // cannot invokeinterface U.clone (NoSuchMethodError). Object.clone would work here, but
-      |  // not in the example in C2 (illegal access to protected). T.clone works in all cases and
-      |  // resolves correctly.
-      |  def f2 = (this: U).clone()
-      |
-      |  // invokevirtual C1.clone()
-      |  def f3 = (this: C1).clone()
-      |}
-      |
-      |class C2 {
-      |  def f1(t: T) = t.clone()  // invokeinterface T.clone
-      |  def f2(t: U) = t.clone()  // invokeinterface T.clone -- Object.clone would be illegal (protected, explained in C1)
-      |  def f3(t: C1) = t.clone() // invokevirtual C1.clone -- Object.clone would be illegal
-      |}
+    s"""trait T { override def clone(): $cloneType = "hi" }
+        |trait U extends T
+        |class C1 extends U with Cloneable {
+        |  // The comments below are true when $cloneType is Object.
+        |  // C1 gets a forwarder for clone that invokes T.clone. this is needed because JVM method
+        |  // resolution always prefers class members, so it would resolve to Object.clone, even if
+        |  // C1 is a subtype of the interface T which has an overriding default method for clone.
+        |
+        |  // invokeinterface T.clone
+        |  def f1 = (this: T).clone()
+        |
+        |  // cannot invokeinterface U.clone (NoSuchMethodError). Object.clone would work here, but
+        |  // not in the example in C2 (illegal access to protected). T.clone works in all cases and
+        |  // resolves correctly.
+        |  def f2 = (this: U).clone()
+        |
+        |  // invokevirtual C1.clone()
+        |  def f3 = (this: C1).clone()
+        |}
+        |
+        |class C2 {
+        |  def f1(t: T) = t.clone()  // invokeinterface T.clone
+        |  def f2(t: U) = t.clone()  // invokeinterface T.clone -- Object.clone would be illegal (protected, explained in C1)
+        |  def f3(t: C1) = t.clone() // invokevirtual C1.clone -- Object.clone would be illegal
+        |}
     """.stripMargin
 
   val runCode =
